Linux build fix: Use int64_t instead of long long in example (#121)
Replace 'long long' with 'int64_t' to match MetadataValue variant type
definition. Fixes template instantiation errors on Linux.

+++ examples/mcp/mcp_example_server.cc
@@ -251,13 +251,13 @@ CallToolResult executeSampleTool(const std::string& name,
                          : "";
     double a = holds_alternative<double>(a_it->second)
                    ? get<double>(a_it->second)
-                   : (holds_alternative<long long>(a_it->second)
-                          ? static_cast<double>(get<long long>(a_it->second))
+                   : (holds_alternative<int64_t>(a_it->second)
+                          ? static_cast<double>(get<int64_t>(a_it->second))
                           : 0.0);
     double b = holds_alternative<double>(b_it->second)
                    ? get<double>(b_it->second)
-                   : (holds_alternative<long long>(b_it->second)
-                          ? static_cast<double>(get<long long>(b_it->second))
+                   : (holds_alternative<int64_t>(b_it->second)
+                          ? static_cast<double>(get<int64_t>(b_it->second))
                           : 0.0);
 
     double calc_result = 0;
